Fix screen lock detection
If the device screen went off by itself, i.e. not because the user pressed
the hardware power button, the screen lock was not immediately (I think it
can be customized in settings). As a result, debugging was not disabled
automatically despite disable_on_lock being set.

Now the lock state is checked every 5 seconds by the DebugStatusService.